# Bite the bullet: make BlockIO async