BlockIO is now an async trait. This was mostly a rote process of adding async to every fn, and adding await everywhere, with the exception of a few changes that should be **closely** reviewed:

The first pass at this commit *only* changed the BlockIO trait, and saw errors where BlockReq's Sender was being closed without a message being sent. This caused BlockReqWaiter's `recv` to return an error.

In order to fix this, we shouldn't pass the Sender (and clones) around anymore.  Several Upstairs functions were changed to *consume* BlockReq, and make sure that a send is sent before the BlockReq is dropped. In order to make sure that other code was not accessing the fields of these structs incorrectly, they were moved to a new module `block_req`.

Another change here: sending through BlockReq, or receiving through BlockReqWaiter consumes the object. A channel of buffer size 1 is created, and it doesn't make sense that users could send or receive multiple messages through these structs.

Have a look at the changes to process_new_io in upstairs/src/lib.rs first, this is where we were previously accessing BlockReq fields directly and handing out clones of Sender. Now we can lean on Rust's borrow checker and make a little more sure that BlockReq and BlockReqWaiter are being used correctly.

If everything is properly async, block_on and block_in_place should no longer be required. If you see block_in_place now, it should be scrutinized and is probably wrong. Likewise, if you're using std::sync::Mutex or other blocking implementations from std, they should be changed to tokio equivalents.

One place where block_in_place is still required is CruciblePseudoFile, which implements std::io::{Read, Write, Seek}, which are *not* async.

Any std::thread::sleep use is a now a bug!

- extent_from_offset no longer returns Result, because it didn't needto

- block_wait was renamed to wait

- FileBlockIO can use reqwest::Client instead of reqwest::blocking::Client, which is nice.

- Volume::add_read_only_parent_create_guest was removed, it wasn't being used

With this commit, the integration tests no longer intermittently hang, and importantly no longer hang even though `worker_threads = 8` was removed. This is a more constrainted environment, where previous to this commit if we boosted the number of threads we would see the hangs go away.

Note I have not yet tested integrating with Propolis and Omicron. That'll happen soon after this commit goes in.