@jmpesp 2022-09-21
I left this running overnight, and unfortunately saw

```
test test::integration_test_volume_write_unwritten_1 has been running for over 60 seconds
```

so it seems the hang investigation isn't over yet.