@jmpesp 2022-09-21
I ran the `measure_iops` tooling to see if it measured any difference:

before:
```
IOPS: [1102.4537, 687.5161, 883.2225, 735.6631, 483.2462, 338.77472, 258.18665, 1143.1948, 627.30444, 456.2048, 342.5279, 240.78601, 1080.2274, 654.8024, 459.7501, 365.50443, 252.71648, 1179.285, 638.8333, 442.99002, 344.06207, 240.87225, 1150.1, 671.9266, 442.46338, 343.79578, 240.9019, 1150.1453, 605.75836, 437.2962, 348.30203, 260.74817, 1106.0356, 627.20276, 452.48642, 354.1651, 235.99773, 1183.6311, 651.46454, 452.3445, 347.12186, 317.9987, 1307.056, 589.8102, 455.087, 348.70407, 444.93475, 1303.6804, 582.1082, 398.1684, 322.53357, 446.92947, 1299.4061, 574.12683, 399.68707, 320.6549, 494.99356, 1282.5076, 589.85187, 414.71634, 321.02795, 503.9869, 1298.0137, 578.8945, 404.2403, 319.4599, 445.59302, 1298.6588, 591.82007, 395.1414, 328.23337, 453.98474, 1322.5967, 553.9318, 403.90668, 321.92664, 432.27814, 1268.5223, 588.0486, 400.45773, 319.73666, 463.78415, 1303.0916, 594.43805, 406.28726, 320.3361, 488.89954, 1234.8977, 582.7446, 400.1917, 317.54135, 467.95303, 1288.9108, 573.6066, 402.38147, 318.74295, 501.7006, 1256.6404, 577.0013, 398.6054]
IOPS mean 603.6118 stddev 337.30182
IOPS min 235.99773 max 1322.5967
BW: [564456.3, 352008.25, 452209.9, 376659.5, 247422.05, 173452.66, 132191.56, 585315.75, 321179.88, 233576.86, 175374.28, 123282.44, 553076.44, 335258.84, 235392.05, 187138.27, 129390.836, 603793.94, 327082.66, 226810.89, 176159.78, 123326.59, 588851.2, 344026.4, 226541.25, 176023.44, 123341.77, 588874.4, 310148.28, 223895.66, 178330.64, 133503.06, 566290.25, 321127.8, 231673.05, 181332.53, 120830.836, 606019.1, 333549.84, 231600.39, 177726.39, 162815.33, 669212.7, 301982.8, 233004.55, 178536.48, 227806.6, 667484.4, 298039.4, 203862.22, 165137.19, 228827.89, 665295.94, 293952.94, 204639.78, 164175.31, 253436.7, 656643.9, 302004.16, 212334.77, 164366.31, 258041.3, 664583.0, 296393.97, 206971.03, 163563.47, 228143.63, 664913.3, 303011.88, 202312.39, 168055.48, 232440.19, 677169.5, 283613.1, 206800.22, 164826.44, 221326.4, 649483.44, 301080.88, 205034.36, 163705.17, 237457.48, 667182.9, 304352.28, 208019.08, 164012.08, 250316.56, 632267.6, 298365.25, 204898.16, 162581.17, 239591.95, 659922.3, 293686.6, 206019.31, 163196.39, 256870.7, 643399.9, 295424.66, 204085.97]
BW mean 309049.25 stddev 172698.53
BW min 120830.836 max 677169.5
real 135.25
user 346.64
sys 15.83
```

after:
```
IOPS: [1145.2971, 638.9745, 950.68506, 788.84454, 515.79095, 399.61655, 270.09067, 1159.2004, 650.4437, 451.4079, 412.898, 267.58505, 1161.3591, 616.88, 494.89865, 402.29095, 247.65329, 1185.6207, 678.83435, 476.32355, 386.71942, 245.40657, 1149.8745, 656.9771, 487.4807, 396.9903, 295.96185, 1070.752, 675.8361, 454.61987, 400.90494, 288.97342, 1145.917, 615.1666, 490.56192, 388.68738, 292.89685, 1263.0776, 677.7369, 504.1715, 389.0397, 447.94522, 1323.9509, 649.6847, 461.3178, 361.76117, 516.95105, 1312.4968, 652.6691, 461.9465, 370.60535, 507.80817, 1357.7465, 640.7908, 451.21896, 340.2732, 513.7798, 1298.5195, 633.3233, 450.4259, 362.17343, 475.78543, 1345.8156, 657.3422, 448.29074, 361.78125, 554.98486, 1349.9246, 627.0067, 436.9413, 352.25555, 478.99664, 1347.14, 633.7543, 449.52478, 361.30673, 482.93488, 1264.2683, 667.7689, 465.84363, 365.88394, 392.85434, 1366.7778, 677.0315, 465.5952, 366.8271, 549.9081, 1266.8243, 652.97687, 453.82428, 364.92722, 558.65265, 1320.8761, 546.8262, 456.21576, 367.81235, 560.4613, 1309.85, 622.8308, 449.36227]
IOPS mean 641.85815 stddev 334.1369
IOPS min 245.40657 max 1366.7778
BW: [586392.1, 327154.94, 486750.75, 403888.4, 264084.97, 204603.67, 138286.42, 593510.6, 333027.2, 231120.84, 211403.78, 137003.55, 594615.9, 315842.56, 253388.11, 205972.97, 126798.484, 607037.8, 347563.2, 243877.66, 198000.34, 125648.164, 588735.75, 336372.28, 249590.13, 203259.03, 151532.47, 548225.0, 346028.1, 232765.38, 205263.33, 147954.39, 586709.5, 314965.3, 251167.7, 199007.94, 149963.19, 646695.75, 347001.28, 258135.81, 199188.33, 229347.95, 677862.9, 332638.56, 236194.72, 185221.72, 264678.94, 671998.4, 334166.6, 236516.61, 189749.94, 259997.78, 695166.2, 328084.88, 231024.11, 174219.88, 263055.25, 664842.0, 324261.53, 230618.06, 185432.8, 243602.14, 689057.56, 336559.22, 229524.86, 185232.0, 284152.25, 691161.4, 321027.44, 223713.95, 180354.84, 245246.28, 689735.7, 324482.2, 230156.69, 184989.05, 247262.66, 647305.4, 341897.7, 238511.94, 187332.58, 201141.42, 699790.25, 346640.13, 238384.75, 187815.47, 281552.94, 648614.06, 334324.16, 232358.03, 186842.73, 286030.16, 676288.56, 279975.0, 233582.47, 188319.92, 286956.2, 670643.2, 318889.38, 230073.48]
BW mean 328631.38 stddev 171078.1
BW min 125648.164 max 699790.25
real 139.63
user 291.05
sys 12.56
```

running both IOPs and BW measurements through ministat:

IOPs:
```
+-------------------------------------------------------------------------------------------------------+
|        x      x   +                                                                                   |
|        x x    x   +x                                                                                  |
|        x x +  x   **          xx                                                               x      |
|x       x x++ +*   ** +  +     xx     + +                                                       x      |
|x x  + xx x++ +*   ** ++ +   + xx   +++ +                                         *          +  x   +  |
|x+x+ + xx x++ +*  x***+*x+  ++ xx +*+++ +                                     x   *+ x       +  x++ +  |
|x+x+ + xx**** *** *******+  +*xxxx+*****+x   x    +       x     +          +x x   *+ x+   x x*xx*** +++|
|   |__|_____________M____M_______A___A__________________________|__|                                   |
+-------------------------------------------------------------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x 100     235.99773     1322.5967      459.7501     603.61181     337.30183
+ 100     245.40657     1366.7778     507.80817      641.8582     334.13689
No difference proven at 95.0% confidence
```

BW:
```
x before
+ after
+-------------------------------------------------------------------------------------------------------+
|        x      x   +                                                                                   |
|        x x    x   +x                                                                                  |
|        x x +  x   **          xx                                                               x      |
|x       x x++ +*   ** +  +     xx     + +                                                       x      |
|x x  + xx x++ +*   ** ++ +   + xx   +++ +                                         *          +  x   +  |
|x+x+ + xx x++ +*  x***+*x+  ++ xx +*+++ +                                     x   *+ x       +  x++ +  |
|x+x+ + xx**** *** *******+  +*xxxx+*****+x   x    +       x     +          +x x   *+ x+   x x*xx*** +++|
|   |__|_____________M____M_______A___A__________________________|__|                                   |
+-------------------------------------------------------------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x 100     120830.84      677169.5     235392.05     309049.25     172698.54
+ 100     125648.16     699790.25     259997.78      328631.4     171078.09
No difference proven at 95.0% confidence
```