@jmpesp 2022-09-21
I ran some tests with crudd and there might be a slight performance increase for reads.

for a crudd-based read test, I 

- initialized three blank 1GB downstairs
- wrote zeroes to the whole lba
- then did two warm up reads of the whole lba, then measured four reads of the whole lba

main (d2be3d81):
```
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rb
1020MiB [66.1MiB/s]

+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rb
1020MiB [65.4MiB/s]

+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rb
1020MiB [69.0MiB/s]

+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rb
1020MiB [67.8MiB/s]
```

PR:
```
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rbp
1020MiB [70.0MiB/s]

+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rbp
1020MiB [70.1MiB/s]

+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rbp
1020MiB [68.7MiB/s]

+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 read
+ pv -rbp
1020MiB [69.8MiB/s]
```

ministat on those MiB/s numbers might show a slight improvement, but the +/- is too large to be sure.

```
x before
+ after
+-------------------------------------------------------------------------------------------------------+
|x              x                                  x                  +     x                +   + +    |
| |_________________________________A______________M__________________|     |_____________A______M_____||
+-------------------------------------------------------------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x   4          65.4            69          67.8        67.075     1.6317169
+   4          68.7          70.1            70         69.65    0.64549722
Difference at 95.0% confidence
	2.575 +/- 2.14694
	3.83899% +/- 3.20081%
	(Student's t, pooled s = 1.2408)
```

for a crudd-based write test, I similarly:

- initialized three blank 1GB downstairs
- did two warm up writes of the whole lba, then measured four writes of the whole lba

main (d2be3d81):
```
+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [45.6MiB/s]

+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [47.8MiB/s]

+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [45.9MiB/s]

+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [46.7MiB/s]
```

this PR:
```
+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [46.3MiB/s]

+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [46.3MiB/s]

+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [46.4MiB/s]

+ pv -rb /dev/zero
+ cargo run --release -p crudd -- --key HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc= --byte-offset 0 --gen 1 -t 127.0.0.1:44101 -t 127.0.0.1:44102 -t 127.0.0.1:44103 write
1020MiB [47.1MiB/s]
```

ministat on those MiB/s numbers shows no change:

```
x before
+ after
+-------------------------------------------------------------------------------------------------------+
|                                   +                                                                   |
|    x            x                 +   +             x                 +                              x|
||___________________________|__________M____AA_______M________|_________________________|              |
+-------------------------------------------------------------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x   4          45.6          47.8          46.7          46.5    0.98319208
+   4          46.3          47.1          46.4        46.525    0.38622101
No difference proven at 95.0% confidence
```